# Removes regex from extended length path check